requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.1
Werkzeug==2.2.2